# go/ogen/gen-gin/go/model_pet.go
/*
 * Pet store schema
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Pet struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name"`

	PhotoUrls []string `json:"photoUrls,omitempty"`

	Status PetStatus `json:"status,omitempty"`
}