# go/ogen/gen-gin/go/model_pet_status.go
/*
 * Pet store schema
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi
// PetStatus : pet status in the store
type PetStatus string

// List of PetStatus
const (
	AVAILABLE PetStatus = "available"
	PENDING PetStatus = "pending"
	SOLD PetStatus = "sold"
)